Add `ttl` method.
In concert with the expiration setting behavior, ttl is expressed in
seconds.

To make the results easier to work with in the generic case, the TTL is
expressed in an integral number of seconds.  It also pessimistically
projects eviction at the _end_ of the second where it will disappear.

Negative values from the Redis PTTL command (notable non-existent keys,
and keys which never expire) are returned as 0 for the same ease of use
considerations. I reckon care ought to be taken to know with which kind
of keys one is dealing.